fix: allow for '.' in file name
Previously it was assumed that file names wouldn't contain any `.`
characters besides the one before the `.go` extension. As I found out
when using the code generated by `99designs/gqlgen`, which by default
outputs generated code to `./graph/shema.resolvers.go`, we actually need
to account for this.

This update enables coverage output for such files to be parsed.